Refactor nzbToMediaConfig -> configuration
